Time: 8 ms (88.75%), Space: 17.4 MB (19.97%) - LeetHub

+++ 662-maximum-width-of-binary-tree/662-maximum-width-of-binary-tree.cpp
@@ -24,14 +24,14 @@ public:
             int first, last ;
             
             for(int i=0 ; i<size ; ++i){
-                long long int cur_id = nodes.front().second - id ;
+                int cur_id = nodes.front().second - id ;
                 TreeNode *tmp = nodes.front().first ;
                 nodes.pop() ;
                 if(i == 0) first = cur_id ; 
                 if(i == size-1) last = cur_id ;
                 
-                if(tmp->left) nodes.push({tmp->left, 2*cur_id+1}) ;
-                if(tmp->right) nodes.push({tmp->right, 2*cur_id+2}) ;
+                if(tmp->left) nodes.push({tmp->left, 2LL*cur_id+1}) ;
+                if(tmp->right) nodes.push({tmp->right, 2LL*cur_id+2}) ;
             }
             ans = max(ans, last-first+1) ;
         }
